refactor(browser): move browser-domain tests into browser_tests
